+++ bigmler/defaults.py
@@ -119,7 +119,10 @@ FLAGS = {'BigMLer': [{'flag': 'debug', 'type': 'boolean'},
                      {'flag': 'batch_prediction_attributes',
                       'type': 'string'},
                      {'flag': 'batch_prediction_tag', 'type': 'string'},
-                     {'flag': 'balance', 'type': 'boolean'}]}
+                     {'flag': 'balance', 'type': 'boolean'},
+                     {'flag': 'weight_field', 'type': 'string'},
+                     {'flag': 'objective_weights', 'type': 'string'},
+                     {'flag': 'label_aggregates', 'type': 'string'}]}
 
 
 def get_user_defaults(defaults_file=DEFAULTS_FILE):